Switch to using new GenerateOptions type from CreateOptions module
