+++ Sources/CreateAPI/Generator/Generator+Paths.swift
@@ -2,6 +2,7 @@
 //
 // Copyright (c) 2021-2022 Alexander Grebenyuk (github.com/kean).
 
+import CreateOptions
 import OpenAPIKit30
 import Foundation
 import GrammaticalNumber
