+++ Sources/CreateAPI/Generator/Generator.swift
@@ -2,6 +2,7 @@
 //
 // Copyright (c) 2021-2022 Alexander Grebenyuk (github.com/kean).
 
+import CreateOptions
 import OpenAPIKit30
 import Foundation
 import GrammaticalNumber
@@ -87,16 +88,7 @@ final class Generator {
     }
     
     var fileHeader: String {
-        var output = ""
-        
-        if let comment = options.fileHeaderComment {
-            output += comment
-        } else {
-            output += """
-            // Generated by Create API
-            // https://github.com/CreateAPI/CreateAPI
-            """
-        }
+        var output = options.fileHeaderComment
         
         if options.isSwiftLintDisabled {
             output += "\n"
